Move scheduled builds to a build scheduler service
