+++ docroot/modules/custom/va_gov_build_trigger/src/Service/BuildRequester.php
@@ -14,7 +14,6 @@ use Drupal\Core\State\StateInterface;
 class BuildRequester implements BuildRequesterInterface {
 
   public const VA_GOV_FRONTEND_VERSION = 'va_gov_build_trigger.frontend_version';
-  public const VA_GOV_LAST_SCHEDULED_BUILD_REQUEST = 'va_gov_build_trigger.last_scheduled_build_request';
 
   /**
    * The content release queue entity.
@@ -85,26 +84,6 @@ class BuildRequester implements BuildRequesterInterface {
    */
   public function resetFrontendVersion() : void {
     $this->state->delete(self::VA_GOV_FRONTEND_VERSION);
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  public function checkScheduledBuild() : void {
-    $currentTime = $this->time->getCurrentTime();
-    $day_of_week = $this->dateFormatter->format($currentTime, 'custom', 'w', 'America/New_York');
-    $hour_of_day = $this->dateFormatter->format($currentTime, 'custom', 'G', 'America/New_York');
-    $last_scheduled_build = $this->state->get(self::VA_GOV_LAST_SCHEDULED_BUILD_REQUEST, 0);
-    $time_since_last_build = ($currentTime - $last_scheduled_build);
-
-    $is_business_day = (1 <= $day_of_week && $day_of_week <= 5);
-    $is_business_hour = (9 <= $hour_of_day && $hour_of_day <= 17);
-
-    if ($is_business_day && $is_business_hour && ($time_since_last_build >= 3600)) {
-      $this->requestFrontendBuild('Scheduled hourly build');
-    }
-
-    $this->state->set(self::VA_GOV_LAST_SCHEDULED_BUILD_REQUEST, $currentTime);
   }
 
 }
