+++ docroot/modules/custom/va_gov_build_trigger/src/Service/BuildRequesterInterface.php
@@ -28,12 +28,4 @@ interface BuildRequesterInterface {
    */
   public function resetFrontendVersion() : void;
 
-  /**
-   * Request a scheduled build if appropriate.
-   *
-   * This method will do nothing if the following criteria are not met:
-   *   * It is currently during business hours (ET)
-   *   * The last scheduled build was more than 60 minutes ago.
-   */
-  public function checkScheduledBuild() : void;
 }
